# src/api-spec/ontology.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Tellus Ontology Engine \u2014 Full API",
    "description": "Comprehensive API specification for the Tellus Ontology Engine. Covers ontology CRUD, object types, properties, link types, and action types.",
    "version": "2.0.0",
    "contact": {
      "name": "Tellus Engineering",
      "email": "engineering@tellus.dev"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v2/ontologies/{ontologyId}/groups": {
      "get": {
        "tags": [
          "Groups"
        ],
        "summary": "List object type groups",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of groups"
          }
        }
      },
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Create a group",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/groups/{groupId}": {
      "put": {
        "tags": [
          "Groups"
        ],
        "summary": "Update a group",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Delete a group",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/groups/{groupId}/members": {
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Add member to group",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Member added"
          }
        }
      },
      "delete": {
        "tags": [
          "Groups"
        ],
        "summary": "Remove member from group",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
...
    "/api/v2/ontologies/{ontologyId}/functions": {
      "get": {
        "tags": [
          "Functions"
        ],
        "summary": "List functions",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of functions"
          }
        }
      },
      "post": {
        "tags": [
          "Functions"
        ],
        "summary": "Register a function",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "apiName",
                  "displayName"
                ],
                "properties": {
                  "apiName": {
                    "type": "string"
                  },
                  "displayName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "inputSchema": {
                    "type": "object"
                  },
                  "outputSchema": {
                    "type": "object"
                  },
                  "codeRepositoryUrl": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Function registered"
          }
        }
      }
    },
    "/api/v2/ontologies/{ontologyId}/functions/{apiName}": {
      "get": {
        "tags": [
          "Functions"
        ],
        "summary": "Get a function by apiName",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Function details"
          }
        }
      },
      "put": {
        "tags": [
          "Functions"
        ],
        "summary": "Update a function",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Function updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Functions"
        ],
        "summary": "Delete a function",
        "parameters": [
          {
            "name": "ontologyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "apiName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted successfully"
          }
        }
      }
    },
...
  },
...
}